test pcn

+++ experiments/inpainting.py
@@ -180,7 +180,7 @@ gibbs_kernel = jax.jit(partial(gibbs_kernel, ts=ts, fwd_sampler=fwd_sampler, sde
                                nparticles=nparticles, transition_sampler=transition_sampler,
                                transition_logpdf=transition_logpdf, likelihood_logpdf=likelihood_logpdf,
                                marg_y=args.marg, explicit_backward=True if args.method == 'gibbs-eb' else False))
-delta = None if len(args.method.split('-')) == 1 else int(args.method.split('-')[-1])
+delta = None if len(args.method.split('-')) == 1 else float(args.method.split('-')[-1])
 pmcmc_kernel = jax.jit(partial(pmcmc_kernel, ts=ts, fwd_ys_sampler=fwd_ys_sampler, sde=sde,
                                ref_sampler=ref_sampler, ref_logpdf=ref_logpdf, transition_sampler=transition_sampler,
                                likelihood_logpdf=likelihood_logpdf, resampling=stratified,
